backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas
numpy
python-dotenv==1.0.1
anthropic==0.34.0
e2b-code-interpreter>=2.0.0
werkzeug==3.0.1
gunicorn==22.0.0
pytest==8.3.2
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.0
scikit-learn
imblearn
xgboost>=2.0.0